pyproject.toml
[project]
name = "talent-scout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
